+++ packages/common-cosmos/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.1.0] - 2023-10-20
+### Changed
+- Version bump with `common` 3.1.3
+
 ## [3.0.3] - 2023-10-17
 ### Changed
 - Update type names to be consistent with main SDK (#189)
@@ -95,7 +99,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix chainTypes not being in deployments
 
 ## [0.0.5] - 2022-06-15
-[Unreleased]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/3.0.3...HEAD
+[Unreleased]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/3.1.0...HEAD
+[3.1.0]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/3.0.3...common-cosmos/3.1.0
 [3.0.3]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/3.0.2...common-cosmos/3.0.3
 [3.0.2]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/3.0.1...common-cosmos/3.0.2
 [3.0.1]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/3.0.0...common-cosmos/3.0.1